+++ tests/test_chainanything.py
@@ -1,4 +1,4 @@
-from recurtools import chainanything
+from recurtools import starchain
 
 
 def test_splitstring():
@@ -7,7 +7,7 @@ def test_splitstring():
     c = 5
     d = [4, 5, 23, 11, 5]
     e = ["dg", "kuku"]
-    assert [x for x in chainanything(a,b,c,d,e, recursive=False, dontflatten=None)] == [1, 2, 3, "d","e", 5, 4, 5, 23, 11, 5, "dg", "kuku"]  # noqa: C416, E501
+    assert [x for x in starchain(a,b,c,d,e, recursive=False, dontflatten=None)] == [1, 2, 3, "d","e", 5, 4, 5, 23, 11, 5, "dg", "kuku"]  # noqa: C416, E501
 
 def test_preservestring():
     # https://stackoverflow.com/questions/72288401/how-to-concat-lists-integers-and-strings-into-one-string/72288721#72288721
@@ -16,7 +16,7 @@ def test_preservestring():
     c = 5
     d = [4, 5, 23, 11, 5]
     e = ["dg", "kuku"]
-    assert [x for x in chainanything(a,b,c,d,e)] == [1, 2, 3, "de", 5, 4, 5, 23, 11, 5, "dg", "kuku"]  # noqa: C416
+    assert [x for x in starchain(a,b,c,d,e)] == [1, 2, 3, "de", 5, 4, 5, 23, 11, 5, "dg", "kuku"]  # noqa: C416
 
 def test_flatten():
     a = [1, 2, 3]
@@ -24,13 +24,13 @@ def test_flatten():
     c = 5
     d = [4, 5, 23, 11, 5]
     e = ["dg", "kuku"]
-    assert [x for x in chainanything(a,b,c,d,e, recursive=True, dontflatten=None)] == [1, 2, 3, "d","e", 5, 4, 5, 23, 11, 5, "d","g", "k","u","k","u"]  # noqa: C416, E501
+    assert [x for x in starchain(a,b,c,d,e, recursive=True, dontflatten=None)] == [1, 2, 3, "d","e", 5, 4, 5, 23, 11, 5, "d","g", "k","u","k","u"]  # noqa: C416, E501
 
 def test_docstring():
     a = "abc"
     b = ["ab","cd"]
-    assert [x for x in chainanything(a, dontflatten=None, recursive=False)] == ["a","b","c"]  # noqa: C416
-    assert [x for x in chainanything(b, dontflatten=None, recursive=False)] == ["ab","cd"]  # noqa: C416
+    assert [x for x in starchain(a, dontflatten=None, recursive=False)] == ["a","b","c"]  # noqa: C416
+    assert [x for x in starchain(b, dontflatten=None, recursive=False)] == ["ab","cd"]  # noqa: C416
 
 def test_join():
     a = [1, 2, 3]
@@ -38,4 +38,4 @@ def test_join():
     c = 5
     d = [4, 5, 23, 11, 5]
     e = ["dg", "kuku"]
-    assert "".join(map(str,chainanything(a,b,c,d,e))) == "123de54523115dgkuku"
+    assert "".join(map(str,starchain(a,b,c,d,e))) == "123de54523115dgkuku"