+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog: recurtools
 
+- rename chainanything to starchain
+
 ## v0.4.0 - update to flatten (BREAKING)
 - **new keyword-only argument** `dontflatten` (optional): type which will not be flattened. Default: `(str, bytes)`.
 - `preservestrings` argument **removed**.